+++ Assets/Xsolla/CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
+## [2.2.0] - 2023-11-29
+### Added
+- Fullscreen mode for Xsolla built-in browser
+
+### Changed
+- `LinkDeviceToAccount` and `AuthViaDeviceID` SDK methods. Methods are simplified, the `DeviceInfo` parameter is not required
+
+### Fixed
+- `AuthViaDeviceID` SDK method backward compatibility for migration from 1.5.x version
+
 ## [2.1.0] - 2023-09-13
 ### Added
 - Centrifugo integration